Move dump_parameters.py up to megatron/core/distributed/
The helper isn't specific to Megatron-FSDP — it operates on any
DTensor-backed optimizer parameter (PyTorch FSDP2, custom DDP, etc.).
The M-FSDP-specific attribute lookups (`megatron_fsdp_param_name`,
`is_qkv`) use getattr with defaults and no-op for non-M-FSDP params.

Move the module from
  megatron/core/distributed/fsdp/src/megatron_fsdp/dump_parameters.py
to
  megatron/core/distributed/dump_parameters.py

Updates the import in `megatron/training/training.py` and the test
file's import + docstring. Test file stays under
`tests/unit_tests/distributed/megatron_fsdp/` because it relies on the
`distributed_setup` fixture in that directory's conftest.py.
