+++ megatron/training/training.py
@@ -52,9 +52,7 @@ from megatron.core.distributed import (
 from megatron.core.distributed.fsdp.mcore_fsdp_adapter import (
     FullyShardedDataParallel as megatron_FSDP,
 )
-from megatron.core.distributed.fsdp.src.megatron_fsdp.dump_parameters import (
-    dump_optimizer_parameters,
-)
+from megatron.core.distributed.dump_parameters import dump_optimizer_parameters
 from megatron.core.enums import ModelType
 from megatron.core.fp8_utils import correct_amax_history_if_needed
 from megatron.core.full_cuda_graph import FullCudaGraphWrapper
@@ -2066,7 +2064,7 @@ def setup_model_and_optimizer(
         # Optional: dump per-rank optimizer parameter sharding for fixture capture.
         # Gated by `DUMP_OPTIMIZER_PARAMETERS_OUTPUT` env var; each rank writes a
         # JSON snapshot under that path. Optimizer-agnostic — fires regardless
-        # of optimizer class. See `megatron_fsdp.dump_parameters` for the schema.
+        # of optimizer class. See `megatron.core.distributed.dump_parameters` for the schema.
         dump_path = os.environ.get("DUMP_OPTIMIZER_PARAMETERS_OUTPUT")
         if dump_path:
             # `MegatronOptimizer` wraps an inner `torch.optim.Optimizer` and
